thousand_words/ca-gcp-developer/requirements.txt
futures==3.0.5
enum34==1.1.6
ply==3.8
future==0.15.2
requests==2.11.0
pyasn1==0.1.9
gax-google-logging-v2==0.8.1
grpc-google-pubsub-v1==0.8.1
simplejson==3.8.2
protobuf==3.0.0
blinker==1.4
poster==0.8.1
gax-google-pubsub-v1==0.8.1
rsa==3.4.2
MarkupSafe==0.23
httplib2==0.9.2
Jinja2==2.8
Flask-DebugToolbar==0.10.0
pyasn1-modules==0.0.8
six==1.10.0
Flask==0.11.1
gcloud==0.18.1
Werkzeug==0.11.10
itsdangerous==0.24
google-api-python-client==1.5.1
googleapis-common-protos==1.2.0
grpc-google-logging-v2==0.8.1
grpcio==1.0.0rc1
GoogleAppEngineCloudStorageClient==1.9.22.1
uritemplate==0.6
google-gax==0.12.5
click==6.6
oauth2client==3.0.0
